fix: make notebook 205 compatible with Jupyter4NFDI hub
This was challenging to fix because the JupyterHub differs quite a bit from the CartoLab Docker environment. I managed to move the code forward so that the notebook now should run in both environments. But I also stumbled across an error in the cell magic with the <code>7z</code> command (TokenError, likely a regression introduced by a new JupyterLab version.

I fixed this by using the %%bash cell magic - this does not show the time, unfortunately. In JupyterHub, this may still produce errors because of missing files or folders - depending on whether the user ran the complete training materials or not. But this is expected. I am fine now with this notebook, if it works with CartoLab and JupyterHub.

For the future, any changes and updates will need to be carefully considered here because this notebook makes use of a lot of bash commands, which will more easily break because notebooks are primarily Python oriented.

I wanted to keep the bash commands because they are significantly shorter and less verbose than the Python alternatives, which is preferable for beginners (and also reduces the length of notebooks and the need to explain all the peculiar Python code around zipping and accessing git version strings etc.).

fixes #29