feat: add citation style override to allow for APA-style references
This was a rabbit hole. The sphinxcontrib-bibtex package does not allow easy modification of reference styles, but I still wanted
to have Dunkel et al. (2025) instead of [Dunkel et al., 2025], which just looks untypical in a scientific context. The override
in custom_bibtex_styles.py fixes this, but does a lot of heavy lifting that may need to be updated at later stages, if sphinxcontrib-bibtex
changes.